[py] move page load strategy methods into Base Options
