+++ py/selenium/webdriver/chromium/options.py
@@ -151,17 +151,6 @@ class ChromiumOptions(ArgOptions):
         else:
             self._arguments = list(set(self._arguments) - args)
 
-    @property
-    def page_load_strategy(self) -> str:
-        return self._caps["pageLoadStrategy"]
-
-    @page_load_strategy.setter
-    def page_load_strategy(self, strategy: str):
-        if strategy in ["normal", "eager", "none"]:
-            self.set_capability("pageLoadStrategy", strategy)
-        else:
-            raise ValueError("Strategy can only be one of the following: normal, eager, none")
-
     def to_capabilities(self) -> dict:
         """
         Creates a capabilities with all the options that have been set
