+++ py/selenium/webdriver/webkitgtk/options.py
@@ -61,17 +61,6 @@ class Options(ArgOptions):
         """
         self._overlay_scrollbars_enabled = value
 
-    @property
-    def page_load_strategy(self):
-        return self._caps["pageLoadStrategy"]
-
-    @page_load_strategy.setter
-    def page_load_strategy(self, strategy):
-        if strategy in ["normal", "eager", "none"]:
-            self.set_capability("pageLoadStrategy", strategy)
-        else:
-            raise ValueError("Strategy can only be one of the following: normal, eager, none")
-
     def to_capabilities(self):
         """
         Creates a capabilities with all the options that have been set and
